requirements.txt
yt-dlp>=2024.1.0
openai-whisper>=20231117
whisperx>=3.1.1
ffmpeg-python>=0.2.0
tqdm>=4.65.0
mcp>=1.0.0
pydantic>=2.0.0
torch>=2.0.0
torchaudio>=2.0.0
# Local diarization dependencies
librosa>=0.10.0
scikit-learn>=1.3.0
scipy>=1.11.0
certifi>=2023.0.0